failed to correctly move one of the data row components to common
