+++ src/pages/view-test-result/cat-b/components/vehicle-details-card/__tests__/vehicle-details-card.spec.ts
@@ -5,7 +5,7 @@ import { ConfigMock } from 'ionic-mocks';
 import { VehicleDetailsCardComponent } from '../vehicle-details-card';
 import { MockComponent } from 'ng-mocks';
 import { DataRowComponent } from '../../../../../../components/common/data-row/data-row';
-import { DataRowCustomComponent } from '../../../../components/data-row-custom/data-row-custom';
+import { DataRowCustomComponent } from '../../../../../../components/common/data-row-custom/data-row-custom';
 
 describe('VehicleDetailsCardComponent', () => {
   let fixture: ComponentFixture<VehicleDetailsCardComponent>;
