+++ src/pages/view-test-result/cat-be/components/business-details-card/__tests__/business-details-card.spec.ts
@@ -4,7 +4,7 @@ import { IonicModule, Config, Platform } from 'ionic-angular';
 import { ConfigMock, PlatformMock } from 'ionic-mocks';
 import { MockComponent } from 'ng-mocks';
 import { DataRowComponent } from '../../../../../../components/common/data-row/data-row';
-import { DataRowCustomComponent } from '../../../../components/data-row-custom/data-row-custom';
+import { DataRowCustomComponent } from '../../../../../../components/common/data-row-custom/data-row-custom';
 import { BusinessDetailsCardComponent } from '../business-details-card';
 import { CatBEUniqueTypes } from '@dvsa/mes-test-schema/categories/BE';
 import { DisplayAddressComponent } from '../../../../../../components/common/display-address/display-address';
