+++ src/pages/view-test-result/cat-be/components/debrief-card/__tests__/debrief-card.spec.ts
@@ -14,7 +14,7 @@ import {
   DrivingFaultsBadgeComponent,
 } from '../../../../../../components/common/driving-faults-badge/driving-faults-badge';
 import { DataRowComponent } from '../../../../../../components/common/data-row/data-row';
-import { DataRowCustomComponent } from '../../../../components/data-row-custom/data-row-custom';
+import { DataRowCustomComponent } from '../../../../../../components/common/data-row-custom/data-row-custom';
 import { DataRowWithListComponent } from '../../../../components/data-row-with-list/data-list-with-row';
 import { FaultsDataRowComponent } from '../../../../components/faults-data-row/faults-data-row';
 import { VehicleChecksDataRowComponent } from '../../../../components/vehicle-checks-data-row/vehicle-checks-data-row';
