+++ src/pages/view-test-result/cat-be/components/vehicle-details-card/__tests__/vehicle-details-card.spec.ts
@@ -5,7 +5,7 @@ import { ConfigMock } from 'ionic-mocks';
 import { VehicleDetailsCardComponent } from '../vehicle-details-card';
 import { MockComponent } from 'ng-mocks';
 import { DataRowComponent } from '../../../../../../components/common/data-row/data-row';
-import { DataRowCustomComponent } from '../../../../components/data-row-custom/data-row-custom';
+import { DataRowCustomComponent } from '../../../../../../components/common/data-row-custom/data-row-custom';
 import { VehicleDetails } from '@dvsa/mes-test-schema/categories/Common';
 import { CatBEUniqueTypes } from '@dvsa/mes-test-schema/categories/BE';
