+++ src/pages/view-test-result/components/contact-details-card/__tests__/contact-details-card.spec.ts
@@ -5,7 +5,7 @@ import { ConfigMock } from 'ionic-mocks';
 import { MockComponent } from 'ng-mocks';
 import { DataRowComponent } from '../../../../../components/common/data-row/data-row';
 import { ContactDetailsCardComponent } from '../contact-details-card';
-import { DataRowCustomComponent } from '../../data-row-custom/data-row-custom';
+import { DataRowCustomComponent } from '../../../../../components/common/data-row-custom/data-row-custom';
 import { CommunicationPreferences, Candidate } from '@dvsa/mes-test-schema/categories/Common';
 import { DisplayAddressComponent } from '../../../../../components/common/display-address/display-address';
