--- src/pages/view-test-result/components/data-row-custom/data-row-custom.html
@@ -1,14 +1,0 @@
-<ion-row [ngClass]="shouldHaveSeperator ? 'mes-data-row-with-separator': 'mes-data-row'">
-    <ion-col col-40>
-      <label>{{ label }}</label>
-      <img
-        alt="Exclamation Indicator"
-        src="assets/imgs/journal/exclamation-indicator.png"
-        class="exclamation-indicator"
-        *ngIf="shouldShowIndicator"
-      />
-    </ion-col>
-    <ion-col col-56>
-      <ng-content></ng-content>
-    </ion-col>
-  </ion-row>
